Dev branch

+++ final-project/db/models/wishlistModel.ts
@@ -26,9 +26,31 @@ export default class WishlistModel {
         {
           $unwind: "$product",
         },
+        {
+          $lookup: {
+            from: "vendors",
+            let: {
+              vendorId: {
+                $toObjectId: "$product.vendorId",
+              },
+            },
+            pipeline: [
+              {
+                $match: {
+                  $expr: {
+                    $eq: ["$_id", "$$vendorId"],
+                  },
+                },
+              },
+            ],
+            as: "vendor",
+          },
+        },
+        {
+          $unwind: "$vendor",
+        },
       ])
       .toArray();
-    // return this.collection().find({ userId }).toArray();
   }
 
   static getWishlistById(id: string) {
@@ -45,13 +67,22 @@ export default class WishlistModel {
   }
 
   static async deleteWishlist(id: string, userId: string) {
-    const result = await this.collection().deleteOne({
-      productId: new ObjectId(id),
+    let result = await this.collection().deleteOne({
+      _id: new ObjectId(id),
       userId: new ObjectId(userId),
     });
+
+    if (result.deletedCount === 0) {
+      result = await this.collection().deleteOne({
+        productId: new ObjectId(id),
+        userId: new ObjectId(userId),
+      });
+    }
+
     if (result.deletedCount === 0) {
       throw new Error("Wishlist not found");
     }
+
     return "Wishlist deleted with ID: " + id;
   }
 }
